test: combine stat selection timing tests
